Refactored marker functionality
