+++ components/MapScreen.tsx
@@ -1,27 +1,11 @@
 import React, { useEffect, useState, useMemo, useCallback, useRef } from "react";
 import { StyleSheet, View, Text } from "react-native";
-import MapView, { Marker, Callout, Region, MapMarker } from "react-native-maps";
+import MapView, { Region, MapMarker } from "react-native-maps";
 import * as Location from "expo-location";
 import { supabase } from "@/lib/supabase";
-import { ThemedText } from "./ThemedText";
-import { useColorScheme } from "@/hooks/useColorScheme";
-import { Ionicons } from '@expo/vector-icons';
-
-// Define the structure of a report
-interface Report {
-  reportid: number;
-  category: string;
-  description: string;
-  location: string; // In the format "(lat,lng)"
-  createdAt: string;
-  imageurl?: string;
-  // Additional fields depending on report type
-  eventtype?: string;
-  hazardtype?: string;
-  itemtype?: string;
-  contactinfo?: string;
-  time?: string;
-}
+import { ReportMarkers } from "./MapScreen/ReportMarkers";
+import { useMarkerClusters } from "./MapScreen/MarkerCluster";
+import { Report, parseLocation } from "./MapScreen/markerUtils";
 
 interface MapScreenProps {
   distanceRadius: number;
@@ -29,58 +13,42 @@ interface MapScreenProps {
   filter?: 'all' | 'hazard' | 'event' | 'lost' | 'found';
 }
 
-// Colors must match `app/forums.tsx` categoryColors
-const FORUM_COLORS = {
-  event: '#7C3AED', // purple-600
-  lost:  '#EAB308', // yellow-500
-  found: '#22C55E', // green-500
-  safety:'#EF4444', // red-500
-};
-
 export default function MapScreen({ distanceRadius, selectedReportId, filter = 'all' }: MapScreenProps) {
   const [location, setLocation] = useState<Location.LocationObject | null>(null);
   const [errorMsg, setErrorMsg] = useState<string | null>(null);
   const [reports, setReports] = useState<Report[]>([]);
   const [selectedReport, setSelectedReport] = useState<Report | null>(null);
-  // track whether a marker's view has finished updating to disable tracksViewChanges
   const markerReadyRef = useRef<Map<number, boolean>>(new Map());
-  const [isLoading, setIsLoading] = useState(true);
-  const colorScheme = useColorScheme() ?? "light";
   const isMountedRef = useRef(true);
   const fetchTimeoutRef = useRef<number | null>(null);
   const mapRef = useRef<MapView>(null);
   const markerRefs = useRef<{ [key: number]: MapMarker | null }>({});
 
-  // Cleanup mounted ref on unmount
   useEffect(() => {
     return () => {
       isMountedRef.current = false;
     };
   }, []);
 
-  // Fetch user location
   useEffect(() => {
     let subscription: Location.LocationSubscription;
 
     (async () => {
       try {
-        // Request location permissions
         let { status } = await Location.requestForegroundPermissionsAsync();
         if (status !== "granted") {
           setErrorMsg("Permission to access location was denied");
           return;
         }
 
-        // Get the current location once
         let currentLocation = await Location.getCurrentPositionAsync({});
         setLocation(currentLocation);
 
-        // Subscribe to location updates
         subscription = await Location.watchPositionAsync(
           {
             accuracy: Location.Accuracy.High,
-            timeInterval: 1000, // update every second
-            distanceInterval: 1, // update every meter
+            timeInterval: 1000,
+            distanceInterval: 1,
           },
           (loc) => {
             setLocation(loc);
@@ -92,7 +60,6 @@ export default function MapScreen({ distanceRadius, selectedReportId, filter = '
       }
     })();
 
-    // Cleanup subscription on unmount
     return () => {
       if (subscription) {
         subscription.remove();
@@ -100,20 +67,17 @@ export default function MapScreen({ distanceRadius, selectedReportId, filter = '
     };
   }, []);
 
-  // Function to fetch all reports
   const fetchReports = useCallback(async () => {
     try {
       console.log("Fetching reports...");
       
-      // Create a timestamp to track this fetch request
       const fetchId = Date.now();
       console.log(`Starting fetch request ${fetchId}`);
 
-      // First fetch basic report data
       const { data: reportsData, error: reportsError } = await supabase
         .from("reports")
         .select("*")
-        .order('createdat', { ascending: false }); // Order by newest first
+        .order('createdat', { ascending: false });
 
       if (reportsError) {
         console.error("Error fetching reports:", reportsError);
@@ -128,7 +92,6 @@ export default function MapScreen({ distanceRadius, selectedReportId, filter = '
 
       console.log(`Found ${reportsData.length} reports for fetch ${fetchId}`);
 
-      // For each report, fetch additional details based on category
       const enhancedReports = await Promise.all(
         reportsData.map(async (report) => {
           try {
@@ -140,7 +103,7 @@ export default function MapScreen({ distanceRadius, selectedReportId, filter = '
                   .from("events")
                   .select("*")
                   .eq("reportid", report.reportid)
-                  .maybeSingle(); // Use maybeSingle instead of single to avoid errors
+                  .maybeSingle();
                 if (eventError) {
                   console.warn(`Error fetching event data for report ${report.reportid}:`, eventError);
                 }
@@ -184,7 +147,6 @@ export default function MapScreen({ distanceRadius, selectedReportId, filter = '
             return { ...report, ...additionalData };
           } catch (reportError) {
             console.error(`Error processing report ${report.reportid}:`, reportError);
-            // Return the basic report data even if additional data fails
             return report;
           }
         })
@@ -192,44 +154,34 @@ export default function MapScreen({ distanceRadius, selectedReportId, filter = '
 
       console.log(`Successfully processed ${enhancedReports.length} reports for fetch ${fetchId}`);
       
-      // Only update state if component is still mounted
       if (isMountedRef.current) {
         setReports(enhancedReports);
       }
     } catch (error) {
       console.error("Error processing reports:", error);
-      // Don't clear existing reports on error, keep what we have
     }
   }, []);
 
-  // Debounced version of fetchReports to prevent too many rapid calls
   const debouncedFetchReports = useCallback(() => {
-    // Clear any existing timeout
     if (fetchTimeoutRef.current) {
       clearTimeout(fetchTimeoutRef.current);
     }
     
-    // Set a new timeout
     fetchTimeoutRef.current = setTimeout(() => {
       if (isMountedRef.current) {
         fetchReports();
       }
-    }, 300); // 300ms debounce
+    }, 300);
   }, [fetchReports]);
 
-  // Fetch reports on component mount and set up real-time subscription
   useEffect(() => {
-    // Initial data fetch
     fetchReports();
 
-    // Set up Supabase subscription for real-time updates
     const setupRealtimeSubscriptions = () => {
       console.log("Setting up real-time subscriptions...");
       
-      // Create a unique channel name to avoid conflicts
       const channelName = `reports-updates-${Date.now()}`;
       
-      // Subscribe to changes in the reports table and related tables
       const channel = supabase
         .channel(channelName)
         .on(
@@ -298,14 +250,12 @@ export default function MapScreen({ distanceRadius, selectedReportId, filter = '
             console.log("Successfully subscribed to real-time updates");
           } else if (status === "CHANNEL_ERROR") {
             console.log("Subscription error, retrying...");
-            // Retry subscription after a delay
             setTimeout(() => {
               setupRealtimeSubscriptions();
             }, 2000);
           }
         });
 
-      // Return cleanup function
       return () => {
         console.log("Cleaning up subscriptions...");
         supabase.removeChannel(channel);
@@ -314,7 +264,6 @@ export default function MapScreen({ distanceRadius, selectedReportId, filter = '
 
     const cleanup = setupRealtimeSubscriptions();
 
-    // Clean up subscription when component unmounts
     return () => {
       isMountedRef.current = false;
       if (fetchTimeoutRef.current) {
@@ -324,24 +273,21 @@ export default function MapScreen({ distanceRadius, selectedReportId, filter = '
     };
   }, [fetchReports, debouncedFetchReports]);
 
-  // Auto-select report when selectedReportId is provided
   useEffect(() => {
     if (selectedReportId && reports.length > 0) {
       const reportToSelect = reports.find(report => report.reportid === selectedReportId);
       if (reportToSelect) {
         setSelectedReport(reportToSelect);
         
-        // Focus the map on the selected report
         const coords = parseLocation(reportToSelect.location);
         if (coords && mapRef.current) {
           mapRef.current.animateToRegion({
             latitude: coords.latitude,
             longitude: coords.longitude,
-            latitudeDelta: 0.005, // Smaller delta for closer zoom
+            latitudeDelta: 0.005,
             longitudeDelta: 0.005,
           }, 100);
           
-          // Show the callout for the selected marker after the animation
           setTimeout(() => {
             const markerRef = markerRefs.current[reportToSelect.reportid];
             if (markerRef) {
@@ -353,151 +299,8 @@ export default function MapScreen({ distanceRadius, selectedReportId, filter = '
     }
   }, [selectedReportId, reports]);
 
-  // Parse location string from the database into latitude and longitude
-  const parseLocation = (
-    locationStr: string
-  ): { latitude: number; longitude: number } | null => {
-    try {
-      // Expected format: "(lat,lng)"
-      const coordsStr = locationStr.substring(1, locationStr.length - 1).trim();
-      const parts = coordsStr.split(",").map(s => parseFloat(s.trim()));
-      if (parts.length < 2) return null;
-      let [lat, lng] = parts;
+  const clusters = useMarkerClusters(reports);
 
-      if (isNaN(lat) || isNaN(lng)) {
-        return null;
-      }
-
-      // Basic validation: lat must be between -90 and 90, lng between -180 and 180.
-      // If values look swapped (e.g., lat outside [-90,90]), swap them.
-      if (Math.abs(lat) > 90 && Math.abs(lng) <= 90) {
-        const tmp = lat;
-        lat = lng;
-        lng = tmp;
-      }
-
-      if (lat < -90 || lat > 90 || lng < -180 || lng > 180) {
-        console.warn('parseLocation: coordinates out of bounds for', locationStr, '->', { lat, lng });
-        return null;
-      }
-
-      return { latitude: lat, longitude: lng };
-    } catch (error) {
-      console.error("Error parsing location:", error);
-      return null;
-    }
-  };
-
-  const deg2rad = (deg: number): number => {
-    return deg * (Math.PI / 180);
-  };
-
-  const getDistanceFromLatLonInKm = (lat1: number, lon1: number, lat2: number, lon2: number): number => {
-    let R = 6371; // Radius of the earth in km
-    let dLat = deg2rad(lat2 - lat1);
-    let dLon = deg2rad(lon2 - lon1);
-    let a =
-      Math.sin(dLat / 2) * Math.sin(dLat / 2) +
-      Math.cos(deg2rad(lat1)) * Math.cos(deg2rad(lat2)) *
-      Math.sin(dLon / 2) * Math.sin(dLon / 2);
-    let c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-    let d = R * c; // Distance in km
-    return d;
-  };
-
-  // Convert a meter offset to approximate degrees latitude/longitude at given latitude
-  const metersToDegreeOffset = (lat: number, meters: number) => {
-    const latDegree = meters / 111320; // approx meters per degree latitude
-    const lngDegree = meters / (111320 * Math.cos((lat * Math.PI) / 180));
-    return { latDegree, lngDegree };
-  };
-
-  // Cluster reports that are within a proximity threshold (meters) to avoid stacking
-  const clusters = useMemo(() => {
-    const thresholdMeters = 2; // distance within which points are considered same cluster
-    const clusters: Array<{ center: { latitude: number; longitude: number }; members: Report[] }> = [];
-
-    const pushToCluster = (r: Report, coords: { latitude: number; longitude: number }) => {
-      for (const c of clusters) {
-        const d = getDistanceFromLatLonInKm(c.center.latitude, c.center.longitude, coords.latitude, coords.longitude) * 1000;
-        if (d <= thresholdMeters) {
-          c.members.push(r);
-          // optionally update cluster center (simple average)
-          const latSum = c.center.latitude * (c.members.length - 1) + coords.latitude;
-          const lngSum = c.center.longitude * (c.members.length - 1) + coords.longitude;
-          c.center.latitude = latSum / c.members.length;
-          c.center.longitude = lngSum / c.members.length;
-          return;
-        }
-      }
-      clusters.push({ center: { latitude: coords.latitude, longitude: coords.longitude }, members: [r] });
-    };
-
-    reports.forEach((r) => {
-      const c = parseLocation(r.location);
-      if (!c) return;
-      pushToCluster(r, c);
-    });
-
-    return clusters;
-  }, [reports]);
-
-  // Given a report and its parsed coords, return a possibly offset display coordinate so markers don't stack
-  const getDisplayCoords = (report: Report, coords: { latitude: number; longitude: number }) => {
-    // find the cluster containing this report
-    const cluster = clusters.find((cl) => cl.members.some((m) => m.reportid === report.reportid));
-    if (!cluster) return coords;
-    const count = cluster.members.length;
-    if (count <= 1) return coords;
-
-    const index = cluster.members.findIndex((m) => m.reportid === report.reportid);
-
-    // Increase radius with cluster size to reduce overlap for larger groups
-    const baseRadius = 6;
-    const radiusMeters = Math.min(baseRadius + count * 2, 40); // cap at 40m
-
-    const angle = (2 * Math.PI * index) / count;
-    const { latDegree, lngDegree } = metersToDegreeOffset(cluster.center.latitude, radiusMeters);
-
-    const adjustedLat = cluster.center.latitude + Math.cos(angle) * latDegree;
-    const adjustedLng = cluster.center.longitude + Math.sin(angle) * lngDegree;
-
-    return { latitude: adjustedLat, longitude: adjustedLng };
-  };
-
-  // Get appropriate marker color based on report category
-  const getMarkerColor = (category: string): string => {
-    switch (category) {
-      case "event":
-        return FORUM_COLORS.event;
-      case "safety":
-        return FORUM_COLORS.safety;
-      case "lost":
-        return FORUM_COLORS.lost;
-      case "found":
-        return FORUM_COLORS.found;
-      default:
-        return "#9E9E9E"; // Gray
-    }
-  };
-
-  // Generate title for the marker based on report type
-  const getReportTitle = (report: Report): string => {
-    switch (report.category) {
-      case "event":
-        return `Event: ${report.eventtype || ""}`;
-      case "safety":
-        return `Hazard: ${report.hazardtype || ""}`;
-      case "lost":
-        return `Lost: ${report.itemtype || ""}`;
-      case "found":
-        return `Found: ${report.itemtype || ""}`;
-      default:
-        return "Report";
-    }
-  };
-
-  // Define the initial map region
   const region: Region | undefined = useMemo(() => {
     if (!location) return undefined;
 
@@ -509,7 +312,6 @@ export default function MapScreen({ distanceRadius, selectedReportId, filter = '
     };
   }, [location]);
 
-  // Loading state
   if (errorMsg) {
     return (
       <View style={[styles.container, styles.centerContent]}>
@@ -536,131 +338,19 @@ export default function MapScreen({ distanceRadius, selectedReportId, filter = '
         followsUserLocation={false}
         showsCompass={false}
       >
-        {/* Render markers for reports matching the selected filter */}
-        {reports.map((report) => {
-          // Map the app filter to report.category values
-          const matchesFilter = (() => {
-            if (filter === 'all') return true;
-            if (filter === 'hazard') return report.category === 'safety';
-            if (filter === 'event') return report.category === 'event';
-            if (filter === 'lost') return report.category === 'lost';
-            if (filter === 'found') return report.category === 'found';
-            return true;
-          })();
-
-          if (!matchesFilter) return null;
-
-          const coords = parseLocation(report.location);
-          if (!coords) return null;
-
-          const distance = getDistanceFromLatLonInKm(
-            location.coords.latitude,
-            location.coords.longitude,
-            coords.latitude,
-            coords.longitude
-          );
-
-          if (distance > distanceRadius) return null;
-
-          // Render a marker using an icon to match the top bar, with circular background and press animation
-          const IconForReport = () => {
-            // Use the same Ionicons names as `app/forums.tsx` for visual consistency
-            switch (report.category) {
-              case 'event':
-                return <Ionicons name="calendar-outline" size={20} color={getMarkerColor(report.category)} />;
-              case 'safety':
-                return <Ionicons name="alert-circle-outline" size={20} color={getMarkerColor(report.category)} />;
-              case 'lost':
-                return <Ionicons name="help-circle-outline" size={18} color={getMarkerColor(report.category)} />;
-              case 'found':
-                return <Ionicons name="checkmark-circle-outline" size={18} color={getMarkerColor(report.category)} />;
-              default:
-                return <Ionicons name="information-circle-outline" size={18} color={getMarkerColor(report.category)} />;
-            }
-          };
-
-          const handlePress = () => {
-            // Toggle selection: if already selected, deselect
-            setSelectedReport((prev) => (prev?.reportid === report.reportid ? null : report));
-          };
-
-          const tracksViewChanges = !markerReadyRef.current.get(report.reportid);
-
-          const displayCoords = getDisplayCoords(report, coords);
-
-          return (
-            <Marker
-              key={report.reportid}
-              ref={(ref) => { markerRefs.current[report.reportid] = ref; }}
-              coordinate={displayCoords}
-              onPress={handlePress}
-              anchor={{ x: 0.5, y: 0.5 }}
-              tracksViewChanges={tracksViewChanges}
-            >
-              <View
-                style={[
-                  styles.iconWrapper,
-                  // If selected, use the marker's category color as background and a slight shadow
-                  selectedReport?.reportid === report.reportid
-                    ? { backgroundColor: getMarkerColor(report.category), shadowColor: getMarkerColor(report.category), elevation: 4 }
-                    : colorScheme === 'dark' ? styles.iconWrapperDark : styles.iconWrapperLight,
-                ]}
-              >
-                {/* If selected, render white icon for contrast */}
-                {selectedReport?.reportid === report.reportid ? (
-                  (() => {
-                    switch (report.category) {
-                      case 'event':
-                        return <Ionicons name="calendar-outline" size={20} color="#fff" />;
-                      case 'safety':
-                        return <Ionicons name="alert-circle-outline" size={20} color="#fff" />;
-                      case 'lost':
-                        return <Ionicons name="help-circle-outline" size={18} color="#fff" />;
-                      case 'found':
-                        return <Ionicons name="checkmark-circle-outline" size={18} color="#fff" />;
-                      default:
-                        return <Ionicons name="information-circle-outline" size={18} color="#fff" />;
-                    }
-                  })()
-                ) : (
-                  IconForReport()
-                )}
-              </View>
-              {/* once rendered, mark ready to stop tracking view changes to stabilize marker */}
-              {tracksViewChanges && (
-                <View
-                  onLayout={() => {
-                    markerReadyRef.current.set(report.reportid, true);
-                  }}
-                  style={{ width: 0, height: 0 }}
-                />
-              )}
-              <Callout tooltip>
-                <View
-                  style={[
-                    styles.callout,
-                    colorScheme === "dark"
-                      ? styles.calloutDark
-                      : styles.calloutLight,
-                  ]}
-                >
-                  <ThemedText type="defaultSemiBold">
-                    {getReportTitle(report)}
-                  </ThemedText>
-                  <ThemedText>{report.description}</ThemedText>
-                  {report.contactinfo && (
-                    <ThemedText>Contact: {report.contactinfo}</ThemedText>
-                  )}
-                  {report.time && (
-                    <ThemedText>
-                      Time: {new Date(report.time).toLocaleString()}
-                    </ThemedText>
-                  )}
-                </View>
-              </Callout>
-            </Marker>
-          );
-        })}
+        <ReportMarkers
+          reports={reports}
+          selectedReport={selectedReport}
+          onReportPress={(report: Report) => {
+            setSelectedReport((prev: Report | null) => (prev?.reportid === report.reportid ? null : report));
+          }}
+          clusters={clusters}
+          markerReadyRef={markerReadyRef}
+          markerRefs={markerRefs}
+          location={location}
+          distanceRadius={distanceRadius}
+          filter={filter}
+        />
       </MapView>
     </View>
   );
@@ -676,17 +366,6 @@ const styles = StyleSheet.create({
   },
   map: {
     flex: 1,
-  },
-  callout: {
-    width: 200,
-    padding: 15,
-    borderRadius: 10,
-  },
-  calloutLight: {
-    backgroundColor: "#fff",
-  },
-  calloutDark: {
-    backgroundColor: "#333",
   },
   errorText: {
     color: "red",
@@ -705,19 +384,5 @@ const styles = StyleSheet.create({
   loadingText: {
     color: "#fff",
     fontSize: 14,
-  },
-  iconWrapper: {
-    width: 44,
-    height: 44,
-    borderRadius: 22,
-    alignItems: 'center',
-    justifyContent: 'center',
-    borderWidth: 0,
-  },
-  iconWrapperLight: {
-    backgroundColor: 'rgba(255,255,255,0.95)',
-  },
-  iconWrapperDark: {
-    backgroundColor: 'rgba(34,34,34,0.95)',
   },
 });
